fix output array assembly for large structures
The result array may become very large. I have an example where it contains 574497 entries. At this point, it's not only slow (what is a different thing to fix) but causes a crash. The maximum number of arguments to give into a function is limited by the browser and is exeeded here (Uncaught RangeError: too many arguments provided for a function call).